chall/requirements.txt
Babel==2.7.0
bcrypt==3.1.7
blinker==1.4
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
Flask==1.0.3
Flask-BabelEx==0.9.3
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
mysqlclient==1.4.2.post1
passlib==1.7.1
pycparser==2.19
pytz==2019.1
PyYAML==5.1.1
six==1.12.0
speaklater==1.3
SQLAlchemy==1.3.5
urllib3==1.25.3
Werkzeug==0.15.4
WTForms==2.2.1
Flask-Session==0.4.0
redis==3.5.3
WTForms-Alchemy==0.18.0